content/html/content/test/test_bug330705-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=330705
-->
<head>
  <title>Test for Bug 330705</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script>
  /* variable is true if the element is focused, false otherwise */
  var inputFocused = false;
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=330705">Mozilla Bug 330705</a>
<p id="display">
  <input onfocus="inputFocused = true" onblur="inputFocused = false" type="text">
  <button></button>
</p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 330705 **/
        SimpleTest.waitForExplicitFinish();
        var isFocused = false;

        function onLoad() {
          document.getElementsByTagName('input')[0].focus();
          document.getElementsByTagName('button')[0].blur();
          ok(inputFocused == true, "the input element is still focused after blur() has been called on the unfocused element");
          SimpleTest.finish();
        }

        addLoadEvent(onLoad);
</script>
</pre>
</body>
</html>
